requirements.txt
langchain
langchain-openai
pydantic
streamlit
faiss-cpu
gtts
python-dotenv
numpy